# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: StudGroup Slice 1 API
  version: 0.2.0
  description: >-
    Contract for the first live vertical slice. Prepared frontend sections are
    intentionally absent until their backend behavior is implemented.
servers:
  - url: https://api.studgroup.example/v1
    description: Production placeholder; deployment supplies the real origin.
security:
  - bearerAuth: []
...
paths:
  /schedule:
    get:
      tags: [Schedule]
      operationId: getSchedule
      summary: Group schedule expanded for the requested date range
      parameters:
        - name: start
          in: query
          required: true
          schema: {type: string, format: date}
        - name: end
          in: query
          required: true
          schema: {type: string, format: date}
      responses:
        '200':
          description: Up to 32 calendar days; unresolved alternating patterns are omitted explicitly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          description: Invalid date range
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/bootstrap:
    post:
      tags: [Session]
      summary: Exchange validated Telegram Mini App data for a StudGroup session
      security: []
      operationId: bootstrapSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionBootstrapRequest'
      responses:
        '200':
          description: Authenticated session, including a possible access gate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionBootstrapResponse'
              examples:
                active:
                  $ref: '#/components/examples/SessionActive'
                noActiveGroup:
                  $ref: '#/components/examples/SessionNoActiveGroup'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Opaque 256-bit-or-stronger session token; never a JWT.
...
  responses:
    Unauthorized:
      description: Telegram bootstrap or bearer session is invalid or expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            sessionExpired:
              value:
                error:
                  code: session_expired
                  message: Сессия истекла. Откройте приложение заново.
                  correlation_id: 01J9Z4A8Z1R9N2Y3P4Q5W6E7T8
                  retryable: false
    Forbidden:
      description: Authenticated user cannot access the active group or resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            membershipSuspended:
              value:
                error:
                  code: membership_suspended
                  message: Не удалось подтвердить участие в группе.
                  correlation_id: 01J9Z4A8Z1R9N2Y3P4Q5W6E7T8
                  retryable: true
...
    RateLimited:
      description: Request rate limit was reached
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  examples:
    SessionActive:
      summary: Active student session
      value:
        access_token: AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
        token_type: Bearer
        expires_at: '2026-10-05T00:00:00Z'
        session:
          user:
            id: 11111111-1111-4111-8111-111111111111
            telegram_user_id: 123456789
            display_name: Анна
            username: anna_student
          access_state: active
          membership:
            id: 22222222-2222-4222-8222-222222222222
            role: student
            status: active
            next_check_at: null
            can_recheck: false
          group:
            id: 33333333-3333-4333-8333-333333333333
            name: ИВТ-21
            timezone: Europe/Moscow
            status: active
            pilot_authorized: true
          permissions: [today.read, homework.read, homework.completion.write]
          server_time: '2026-10-04T12:00:00Z'
    SessionNoActiveGroup:
      summary: Authenticated user without an activated group
      value:
        access_token: BBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBB
        token_type: Bearer
        expires_at: '2026-10-05T00:00:00Z'
        session:
          user:
            id: 11111111-1111-4111-8111-111111111111
            telegram_user_id: 123456789
            display_name: Анна
            username: null
          access_state: no_active_group
          membership: null
          group: null
          permissions: []
          server_time: '2026-10-04T12:00:00Z'
...
  schemas:
    ScheduleResponse:
      type: object
      required: [generated_at, group_timezone, start, end, week_state, first_week_anchor, selected_week, lessons]
      properties:
        generated_at: {type: string, format: date-time}
        group_timezone: {type: string}
        start: {type: string, format: date}
        end: {type: string, format: date}
        week_state: {type: string, enum: [ready, needs_clarification]}
        first_week_anchor: {type: [string, 'null'], format: date}
        selected_week: {type: [string, 'null'], enum: [first, second, null]}
        lessons:
          type: array
          items:
            $ref: '#/components/schemas/LessonOccurrence'
    LessonOccurrence:
      type: object
      required: [id, subject, starts_at, ends_at, teacher, location, status]
      properties:
        id: {type: string}
        subject: {type: string}
        starts_at: {type: string, format: date-time}
        ends_at: {type: string, format: date-time}
        teacher: {type: [string, 'null']}
        location: {type: [string, 'null']}
        status: {type: string, enum: [scheduled]}
    SessionBootstrapRequest:
      type: object
      additionalProperties: false
      required: [init_data]
      properties:
        init_data:
          type: string
          minLength: 1
          description: Raw Telegram Mini App initData string, not parsed client data.
    SessionBootstrapResponse:
      type: object
      additionalProperties: false
      required: [access_token, token_type, expires_at, session]
      properties:
        access_token:
          type: string
          minLength: 43
          maxLength: 43
          pattern: '^[A-Za-z0-9_-]{43}$'
          description: Unpadded base64url encoding of 32 random bytes; never a JWT.
        token_type:
          type: string
          const: Bearer
        expires_at:
          type: string
          format: date-time
        session:
          $ref: '#/components/schemas/SessionContext'
    SessionContext:
      type: object
      additionalProperties: false
      required: [user, access_state, permissions, server_time]
      properties:
        user:
          $ref: '#/components/schemas/User'
        access_state:
          $ref: '#/components/schemas/AccessState'
        membership:
          oneOf:
            - $ref: '#/components/schemas/Membership'
            - type: 'null'
        group:
          oneOf:
            - $ref: '#/components/schemas/Group'
            - type: 'null'
        permissions:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/Permission'
        server_time:
          type: string
          format: date-time
    User:
      type: object
      additionalProperties: false
      required: [id, telegram_user_id, display_name]
      properties:
        id:
          type: string
          format: uuid
        telegram_user_id:
          type: integer
          format: int64
        display_name:
          type: string
        username:
          type: [string, 'null']
    Membership:
      type: object
      additionalProperties: false
      required: [id, role, status, can_recheck]
      properties:
        id:
          type: string
          format: uuid
        role:
          $ref: '#/components/schemas/MemberRole'
        status:
          type: string
          enum: [active, suspended, left]
        next_check_at:
          type: [string, 'null']
          format: date-time
        can_recheck:
          type: boolean
    Group:
      type: object
      additionalProperties: false
      required: [id, name, timezone, status, pilot_authorized]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        timezone:
          type: string
          example: Europe/Moscow
        status:
          type: string
          enum: [active, deletion_pending, read_only]
        pilot_authorized:
          type: boolean
          description: Always true for a group exposed by an active pilot session.
    AccessState:
      type: string
      enum:
        - active
        - no_active_group
        - membership_suspended
        - membership_left
        - group_deletion_pending
    MemberRole:
      type: string
      enum: [student, deputy, headman]
    Permission:
      type: string
      enum:
        - schedule.read
        - today.read
        - homework.read
        - homework.completion.write
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message, correlation_id, retryable]
      properties:
        code:
          type: string
          enum:
            - invalid_request
            - unauthenticated
            - initdata_invalid
            - initdata_expired
            - session_expired
            - no_active_group
            - not_group_member
            - membership_suspended
            - membership_left
            - group_not_authorized_for_pilot
            - invitation_expired
            - invitation_consumed
            - group_deletion_pending
            - permission_denied
            - not_found
            - gone
            - revision_conflict
            - rate_limited
            - service_unavailable
            - internal_error
        message:
          type: string
        correlation_id:
          type: string
        retryable:
          type: boolean
        retry_after_seconds:
          type: [integer, 'null']
          minimum: 1
        field_path:
          type: [string, 'null']
          description: JSON Pointer to an invalid request field when applicable.
        details:
          type: [object, 'null']
          additionalProperties: true